requirements.txt
fastapi
uvicorn
streamlit
requests
bs4
sentence-transformers
ctransformers
numpy
